pyproject.toml
[project]
name = "llm-happiness"
version = "0.1.0"
description = ""
authors = [
    {name = "renatex333",email = "renato9302@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langgraph (>=0.6.8,<0.7.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.35,<0.4.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.3.30,<0.4.0)",
    "langchain-tavily (>=0.2.11,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
